pyproject.toml
[project]
name = "react-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
    "wikipedia>=1.4.0",
]
